Integrate into SparkQueryDispatcher
