+++ async-query-core/src/main/java/org/opensearch/sql/spark/validator/GrammarElement.java
@@ -9,16 +9,16 @@ import lombok.AllArgsConstructor;
 
 @AllArgsConstructor
 enum GrammarElement {
-  ALTER_NAMESPACE("ALTER DATABASE/TABLE/NAMESPACE"),
+  ALTER_NAMESPACE("ALTER (DATABASE|TABLE|NAMESPACE)"),
   ALTER_VIEW("ALTER VIEW"),
-  CREATE_NAMESPACE("CREATE DATABASE/TABLE/NAMESPACE"),
+  CREATE_NAMESPACE("CREATE (DATABASE|TABLE|NAMESPACE)"),
   CREATE_FUNCTION("CREATE FUNCTION"),
   CREATE_VIEW("CREATE VIEW"),
-  DROP_NAMESPACE("DROP DATABASE/TABLE/NAMESPACE"),
+  DROP_NAMESPACE("DROP (DATABASE|TABLE|NAMESPACE)"),
   DROP_FUNCTION("DROP FUNCTION"),
   DROP_VIEW("DROP VIEW"),
   DROP_TABLE("DROP TABLE"),
-  REPAIR_TABLE("REPAIR TABLE"), // does this conflict with DROP_NAMESPACE?
+  REPAIR_TABLE("REPAIR TABLE"),
   TRUNCATE_TABLE("TRUNCATE TABLE"),
   // DML Statements
   INSERT("INSERT"),
@@ -52,7 +52,7 @@ enum GrammarElement {
   ANALYZE_TABLE("ANALYZE TABLE(S)"),
   CACHE_TABLE("CACHE TABLE"),
   CLEAR_CACHE("CLEAR CACHE"),
-  DESCRIBE_NAMESPACE("DESCRIBE (NAMESPACE|DATABASE|SCHEMA"),
+  DESCRIBE_NAMESPACE("DESCRIBE (NAMESPACE|DATABASE|SCHEMA)"),
   DESCRIBE_FUNCTION("DESCRIBE FUNCTION"),
   DESCRIBE_QUERY("DESCRIBE QUERY"),
   DESCRIBE_TABLE("DESCRIBE TABLE"),
