+++ async-query/src/test/java/org/opensearch/sql/spark/asyncquery/AsyncQueryExecutorServiceSpec.java
@@ -102,6 +102,8 @@ import org.opensearch.sql.spark.response.JobExecutionResponseReader;
 import org.opensearch.sql.spark.response.OpenSearchJobExecutionResponseReader;
 import org.opensearch.sql.spark.scheduler.AsyncQueryScheduler;
 import org.opensearch.sql.spark.scheduler.OpenSearchAsyncQueryScheduler;
+import org.opensearch.sql.spark.validator.GrammarElementValidatorFactory;
+import org.opensearch.sql.spark.validator.SQLQueryValidator;
 import org.opensearch.sql.storage.DataSourceFactory;
 import org.opensearch.test.OpenSearchIntegTestCase;
 
@@ -308,6 +310,8 @@ public class AsyncQueryExecutorServiceSpec extends OpenSearchIntegTestCase {
             emrServerlessClientFactory,
             new OpenSearchMetricsService(),
             sparkSubmitParametersBuilderProvider);
+    SQLQueryValidator sqlQueryValidator =
+        new SQLQueryValidator(new GrammarElementValidatorFactory());
     SparkQueryDispatcher sparkQueryDispatcher =
         new SparkQueryDispatcher(
             this.dataSourceService,
@@ -318,7 +322,8 @@ public class AsyncQueryExecutorServiceSpec extends OpenSearchIntegTestCase {
                 sessionConfigSupplier,
                 sessionIdProvider),
             queryHandlerFactory,
-            new DatasourceEmbeddedQueryIdProvider());
+            new DatasourceEmbeddedQueryIdProvider(),
+            sqlQueryValidator);
     return new AsyncQueryExecutorServiceImpl(
         asyncQueryJobMetadataStorageService,
         sparkQueryDispatcher,
